big refactoring for norminette using helper methods
