+++ includes/map_to_edges.h
@@ -23,8 +23,12 @@ void		explore_edge(t_edge_exploration_context *context);
 int			extract_edge_recursively(char **map, t_list **edges);
 int			process_map(t_edge_exploration_context *context, t_list **edges,
 				int height_map, int width_map);
+int			process_row(t_edge_exploration_context *context, t_list **edges,
+				int y, int width_map);				
 void		update_context(t_edge_exploration_context *context, t_point2d coord,
 				t_point2d delta, t_direction dir);
+int			process_direction(t_edge_exploration_context *context,
+				t_list **edges, t_point2d delta, t_direction dir);
 int			add_edge(t_segment_d new_segment, t_list **edges);
 t_direction	**malloc2DArray(char **map);
